Moved to .hpp/.cpp, switched to get_value()
