+++ cpp/include/cudf/lists/detail/utilities.hpp
